basic battle logic completed

+++ GalaxyWars/Species/Species.cs
@@ -12,6 +12,6 @@ namespace GalaxyWars.Species
     public string allowedVehicle { get; set; }
     public string affiliation { get; set; }
     public string homePlanet { get; set; }
-    public int population { get; set; }
+    public double population { get; set; }
   }
 }
